kata/find_the_vowels.cpp

+++ kata/find_the_vowels.cpp
@@ -6,7 +6,7 @@ std::vector<int> vowelIndices(const std::string& word)
     std::vector<int> res;
     char charArray[] = {'a', 'e', 'i', 'o', 'u'};
 
-    for (int i = 0; i < word.size(); ++i)
+    for (int i = 0; i < static_cast<int>(word.size()); ++i)
     {
         for (const auto& c : charArray)
         {
